Add a more visible demo animation, with a moving alpha gradient
A solid box with a gradient inside, that cycles its color and moves the
starting point of the gradient - this is much more visible, and better
demonstrates how the alpha channel can be used.

That visibility also more clearly shows that there is a bug somewhere,
which causes the output to be wrong when it is piped through ffmpeg.
However, raw video works fine, and this does not touch that pipe code.