Fix broken output when piping through ffmpeg (remove bad option)
The visible effect was that the colors and alpha were mixed up in the
resulting video, and that ffmpeg usually gave an error after encoding.
Removing this option from the commandline seems to fix the problem.

I don't really know _why_ this option causes problems (though I have
some suspicions), and the docs didn't give any warnings about it (that
I saw, at least), but at least now it seems to work.